+++ src/MyVideoExplorer/theme/models.py
@@ -16,6 +16,7 @@ class ThemeConfig:
 
     # Color Palette - Backgrounds
     color_background_main: str = "#111111"
+    color_background_section: str = "#161616"
     color_background_gradient: str = "#131613"
     color_surface_primary: str = "rgba(59, 89, 99, 99)"
     color_surface_alternate: str = "#151B21"